feat: POST /clips/upload to POST /clips
